refactor: use init for async stuff (#465)
* refactor: use init for async stuff

* willStart waits for init completion

* refactor: better promise returning and catching

* refactor: async await for init

* better use of await and willStart fixed

* end and destroy use phases

* refactor: new timed media renderer class

* fix: adapt for improved playout engine

+++ src/renderers/BaseRenderer.js
@@ -23,6 +23,8 @@ import PauseBehaviour from '../behaviours/PauseBehaviour';
 
 const SEEK_TIME = 10;
 
+const debugPhase = true;
+
 const getBehaviourEndTime = (behaviour: Object) => {
     if(behaviour.duration !== undefined) {
         const endTime = behaviour.start_time + behaviour.duration;
@@ -32,11 +34,14 @@ const getBehaviourEndTime = (behaviour: Object) => {
 }
 
 export const RENDERER_PHASES = {
+    CONSTRUCTING: 'CONSTRUCTING',
     CONSTRUCTED: 'CONSTRUCTED',
     START: 'START',
     MAIN: 'MAIN',
-    END: 'END',
+    COMPLETING: 'COMPLETING',
     ENDED: 'ENDED',
+    DESTROYED: 'DESTROYED',
+    MEDIA_FINISHED: 'MEDIA_FINISHED', // done all its rendering and ready to move on, but not ended
 };
 
 export default class BaseRenderer extends EventEmitter {
@@ -95,8 +100,6 @@ export default class BaseRenderer extends EventEmitter {
     _choiceBehaviourData: Object;
 
     _linkBehaviour: Object;
-
-    _hasEnded: boolean;
 
     inVariablePanel: boolean;
 
@@ -232,16 +235,30 @@ export default class BaseRenderer extends EventEmitter {
         this._analytics = analytics;
         this.inVariablePanel = false;
         this._preloadedBehaviourAssets = [];
-        this._preloadBehaviourAssets();
-        this._preloadIconAssets();
+        this._preloadBehaviourAssets().catch(e =>
+            logger.warn(e, 'Could not preload behaviour assets'));
+        this._preloadIconAssets().catch(e =>
+            logger.warn(e, 'Could not preload icon assets'));
         this._loopCounter = 0;
-        this.phase = RENDERER_PHASES.CONSTRUCTED;
+        this.phase = RENDERER_PHASES.CONSTRUCTING;
         this._inPauseBehaviourState = false;
     }
 
-    willStart(elementName: ?string, elementId: ?string) {
-        this.inVariablePanel = false;
+    // run any code that may be asynchronous
+    async init() {
+        // eslint-disable-next-line max-len
+        throw new Error('Need to override this class to run async code and set renderer phase to CONSTRUCTED');
+    }
+
+    willStart(elementName: ?string, elementId: ?string): boolean {
+        if (this.phase === RENDERER_PHASES.CONSTRUCTING) {
+            setTimeout(() => this.willStart(elementName, elementId), 100);
+            return false;
+        }
+        // eslint-disable-next-line max-len
+        if (debugPhase) logger.info('PHASE will starting', this._representation.name, this.phase);        
         this.phase = RENDERER_PHASES.START;
+        this.inVariablePanel = false;
 
         this._runStartBehaviours();
 
@@ -251,6 +268,7 @@ export default class BaseRenderer extends EventEmitter {
             const { name, id } = this._representation;
             this._player.addDetails(elementName, elementId, name, id)
         }
+        return true;
     }
 
     _runStartBehaviours() {
@@ -284,9 +302,9 @@ export default class BaseRenderer extends EventEmitter {
      */
 
     start() {
+        if (debugPhase) logger.info('PHASE starting', this._representation.name, this.phase);
         this.phase = RENDERER_PHASES.MAIN;
         this.emit(RendererEvents.STARTED);
-        this._hasEnded = false;
         this._timer.start();
         if (!this._playoutEngine.isPlaying()) {
             this._timer.pause();
@@ -301,10 +319,23 @@ export default class BaseRenderer extends EventEmitter {
         this._runDuringBehaviours();
     }
 
-    end() {
-        this.phase = RENDERER_PHASES.ENDED;
+    end(): boolean {
+        switch (this.phase) {
+        case (RENDERER_PHASES.ENDED):
+        case (RENDERER_PHASES.DESTROYED):
+            // eslint-disable-next-line max-len
+            if (debugPhase) logger.info('PHASE base ended already', this._representation.name, this.phase);
+            return false;
+        default:
+            break;
+        };
+        if (debugPhase) logger.info('PHASE base ending', this._representation.name, this.phase);
         this._player.disconnectScrubBar(this);
-        this._clearBehaviourElements()
+        try{
+            this._clearBehaviourElements()
+        } catch (e) {
+            logger.info(e);
+        }
         this._reapplyLinkConditions();
         clearTimeout(this._linkFadeTimeout);
         this._player.removeListener(PlayerEvents.LINK_CHOSEN, this._handleLinkChoiceEvent);
@@ -318,10 +349,17 @@ export default class BaseRenderer extends EventEmitter {
             this._handlePlayPauseButtonClicked,
         );
         this._lastSetTime = 0;
+        this.phase = RENDERER_PHASES.ENDED;
+        return true;
     }
 
-    hasEnded(): boolean {
-        return this._hasEnded;
+    // has the media finished?
+    hasMediaEnded(): boolean {
+        return (
+            this.phase === RENDERER_PHASES.MEDIA_FINISHED
+            || this.phase === RENDERER_PHASES.ENDED
+            || this.phase === RENDERER_PHASES.DESTROYED
+        );
     }
 
     // if we have any start pauses, complete those behaviours early
@@ -593,8 +631,8 @@ export default class BaseRenderer extends EventEmitter {
     }
 
     complete() {
-        this.phase = RENDERER_PHASES.END;
-        this._hasEnded = true;
+        if (debugPhase) logger.info('PHASE completing', this._representation.name, this.phase);
+        this.phase = RENDERER_PHASES.COMPLETING;
         this._timer.pause();
         if (!this._linkBehaviour ||
             (this._linkBehaviour && !this._linkBehaviour.forceChoice)) {
@@ -627,8 +665,9 @@ export default class BaseRenderer extends EventEmitter {
         this._preloadedBehaviourAssets = [];
         const assetCollectionIds = this._representation.asset_collections.behaviours ?
             this._representation.asset_collections.behaviours : [];
-        assetCollectionIds.forEach((behaviour) => {
-            this._fetchAssetCollection(behaviour.asset_collection_id)
+        return Promise.all(assetCollectionIds.map((behaviour) => {
+        // assetCollectionIds.forEach((behaviour) => {
+            return this._fetchAssetCollection(behaviour.asset_collection_id)
                 .then((assetCollection) => {
                     if (assetCollection.assets.image_src) {
                         return this._fetchMedia(assetCollection.assets.image_src);
@@ -641,8 +680,11 @@ export default class BaseRenderer extends EventEmitter {
                         image.src = imageUrl;
                         this._preloadedBehaviourAssets.push(image);
                     }
+                }).catch((err) => {
+                    logger.error(err,
+                        `could not preload behaviour asset ${behaviour.asset_collection_id}`);
                 });
-        });
+        }));
     }
 
     _preloadIconAssets() {
@@ -671,6 +713,8 @@ export default class BaseRenderer extends EventEmitter {
                         logger.info(`Preloading icon ${imageUrl}`);
                         this._preloadedIconAssets.push(image);
                     }
+                }).catch((err) => {
+                    logger.error(err, `could not preload icon asset ${iconAssetCollection}`);
                 });
         }));
     }
@@ -880,43 +924,46 @@ export default class BaseRenderer extends EventEmitter {
                 return Promise.resolve();
             }
 
-            // build icons
-            const iconSrcPromises = this._getIconSourceUrls(narrativeElementObjects, behaviour);
-
             // find out which link is default
             const defaultLinkId = this._getDefaultLink(narrativeElementObjects);
 
             // go through asset collections and render icons
-            return iconSrcPromises.then((iconObjects) => {
+            return this._getIconSourceUrls(narrativeElementObjects, behaviour)
+                .then((iconObjects) => {
 
-                this._player.clearLinkChoices();
-                iconObjects.forEach((iconSpecObject) => {
-
-                    this._buildLinkIcon(iconSpecObject, behaviourOverlay.overlay);
-                });
-                if (iconObjects.length > 1 || showIfOneLink) {
-                    this._showChoiceIcons({
-                        defaultLinkId, // id for link to highlight at start
-                        forceChoice, // do we highlight
-                        disableControls, // are controls disabled while icons shown
-                        countdown, // do we animate countdown
-                        iconOverlayClass, // css classes to apply to overlay
-
-                        behaviourOverlay,
-                        choiceCount: iconObjects.length,
+                    this._player.clearLinkChoices();
+                    iconObjects.forEach((iconSpecObject) => {
+                        this._buildLinkIcon(iconSpecObject, behaviourOverlay.overlay);
                     });
+                    if (iconObjects.length > 1 || showIfOneLink) {
+                        this._showChoiceIcons({
+                            defaultLinkId, // id for link to highlight at start
+                            forceChoice, // do we highlight
+                            disableControls, // are controls disabled while icons shown
+                            countdown, // do we animate countdown
+                            iconOverlayClass, // css classes to apply to overlay
 
-                    // callback to say behaviour is done, but not if user can
-                    // change their mind
-                    if (!forceChoice) {
+                            behaviourOverlay,
+                            choiceCount: iconObjects.length,
+                        });
+
+                        // callback to say behaviour is done, but not if user can
+                        // change their mind
+                        if (!forceChoice) {
+                            callback();
+                        }
+                    } else {
+                        logger.info('Link Choice behaviour ignored - only one link');
+                        this._linkBehaviour.forceChoice = false;
                         callback();
                     }
-                } else {
-                    logger.info('Link Choice behaviour ignored - only one link');
-                    this._linkBehaviour.forceChoice = false;
+                }).catch((err) => {
+                    logger.error(err, 'could not get assets for rendering link icons');
                     callback();
-                }
-            });
+                });
+        }).catch((err) => {
+            logger.error(err, 'Could not get next steps for rendering links');
+            callback();
         });
     }
 
@@ -1132,15 +1179,18 @@ export default class BaseRenderer extends EventEmitter {
             iconOverlayClass,
             behaviourOverlay,
             choiceCount,
-        );
-        this._player.enableLinkChoiceControl();
-        if (disableControls) {
-            // disable transport controls
-            this._player.disableControls();
-        }
-        if (countdown) {
-            this._player.startChoiceCountdown(this);
-        }
+        ).then(() => {
+            if (disableControls) {
+                // disable transport controls
+                this._player.disableControls();
+            }
+            if (countdown) {
+                this._player.startChoiceCountdown(this);
+            }
+            this._player.enableLinkChoiceControl();
+        }).catch((err) => { // REFACTOR: this returns a promise
+            logger.error(err, 'could not render link choice icons')  ;
+        });
     }
 
     // user has made a choice of link to follow - do it
@@ -1232,6 +1282,8 @@ export default class BaseRenderer extends EventEmitter {
         callback();
     }
 
+    // REFACTOR note: these are called by the behaviour, without knowing what will happen
+    // via behaviour map
     _applyShowImageBehaviour(behaviour: Object, callback: () => mixed) {
         const behaviourAssetCollectionMappingId = behaviour.image;
         const assetCollectionId =
@@ -1249,7 +1301,12 @@ export default class BaseRenderer extends EventEmitter {
                         this._overlayImage(imageUrl, behaviour.id);
                     }
                     callback();
+                })
+                .catch((err) => {
+                    logger.error(err, 'could not get image for show image behaviour');
                 });
+        } else {
+            logger.error('No asset collection id for show image behaviour');
         }
     }
 
@@ -1319,8 +1376,8 @@ export default class BaseRenderer extends EventEmitter {
             this._analytics,
         );
     }
-
     // //////////// end of variables panel choice behaviour
+
     _clearBehaviourElements() {
         const behaviourElements =
             document.querySelectorAll(`[behaviour-renderer="${this._rendererId}"]`);
@@ -1411,7 +1468,16 @@ export default class BaseRenderer extends EventEmitter {
      * @return {void}
      */
     destroy() {
-        this.end();
+        if (debugPhase) logger.info('PHASE destroying', this._representation.name, this.phase);
+        if (this.phase === RENDERER_PHASES.DESTROYED) {
+            // eslint-disable-next-line max-len
+            if (debugPhase) logger.info('PHASE destroying - already destroyed', this._representation.name, this.phase);
+            return false;
+        }
+        if (!this.phase === RENDERER_PHASES.ENDED) {
+            if (debugPhase) logger.info('PHASE destroying need to end first');
+            this.end();
+        }
         this._clearBehaviourElements();
         if (this._behaviourRunner) {
             this._behaviourRunner.destroyBehaviours();
@@ -1419,6 +1485,7 @@ export default class BaseRenderer extends EventEmitter {
         // we didn't find any behaviours to run, so emit completion event
         this.emit(RendererEvents.DESTROYED);
         this._destroyed = true;
+        return true;
     }
 
     isIosPlayoutEngine() {
